Function to run FI-LICA

This function runs FI-LICA.

Usage

FI_LICA(
  data,
  ncomp = 5,
  flica_niter = 1500,
  n = 20,
  flica_niter2 = 1000,
  rescale = TRUE,
  re_completer = NULL
)

Arguments

data

data from different modalities in lists

ncomp

number of components to be used in LICA

flica_niter

number of LICA iterations to be used in FI-LICA's Step 1 (initialization process; completer case analysis)

n

number of FI-LICA updates

flica_niter2

number of LICA iterations to be used in FI-LICA's Step 2 (optimization process)

rescale

rescale H and XW (default is TRUE)

re_completer

results from completer case analysis. Default is NULL and completer case analysis will be run and saved during the process; otherwise, the given results will be used.

Value

The resulting list includes the estimated results, the convergence measures for H and for XW, and the dF history from FI-LICA

Examples

# generate data
data = data_generate_MAR(nsubj = 100, seed = 7452, n_miss = 5)
# data with missing in lists
data = list(data$Y1_miss, data$Y2_miss)
# run FI-LICA
#re_filica = FI_LICA(data = data, ncomp = 5, flica_niter = 1500, n = 20, flica_niter2 = 1000, rescale = TRUE, re_completer = NULL)
# or
#re_completer = LICA(data = data, ncomp = 5, niter = 1500, method = "completer")
#re_filica = FI_LICA(data = data, ncomp = 5, flica_niter = 1500, n = 20, flica_niter2 = 1000, rescale = TRUE, re_completer = re_completer)