Data generation function for MAR (with H from a mixture distribution)

This function generates the simulation data (2 modalities) for the setting of MAR, with one continuous covariate (C1), one binary covariate (C2), and H from a mixture distribution based on C1 and C2

Usage

data_generate_MAR(
  nsubj = 100,
  nvoxel1 = 1000,
  nvoxel2 = 3000,
  true_ncomp = 2,
  seed = 1234,
  n_miss = 5,
  es_C1H1 = 0.5,
  es_C2H2 = 0.5
)

Arguments

nsubj

number of subjects

nvoxel1

number of voxels in modality 1

nvoxel2

number of voxels in modality 2

true_ncomp

number of true component

seed

seed

n_miss

number of missing subjects

es_C1H1

correlation between the first covariate (C1, continuous) and the first component of H (H1)

es_C2H2

cohen's d between the second covariate (C2, binary) and the second component of H (H2)

Value

The resulting dataset has lists including the no-missing data (Y1, Y2), order of subjects who were implemented missing (subj1_miss, subj2_miss), missing data (Y1_miss, Y2_miss), and other intermediate parameters during the data generation process.

Examples

dat = data_generate_MAR(nsubj = 100, seed = 1234, n_miss = 5); names(dat)
#>  [1] "Y1"         "Y2"         "H"          "X1"         "X2"        
#>  [6] "W1"         "W2"         "XW1"        "XW2"        "subj1"     
#> [11] "subj2"      "subj1_miss" "subj2_miss" "subj"       "Y1_miss"   
#> [16] "Y2_miss"    "seed"       "C1"         "C2"         "H1"        
#> [21] "H2"