Function to run LICA on complete cases

This function calls the matlab function to run LICA on complete cases

Usage

flica_completer(ncomp, flica_niter)

Arguments

ncomp

number of components to be used in LICA

flica_niter

number of iterations to be used in LICA

Value

The resulting dataset is the results estimated from LICA

Examples

# Save data at "./MATLAB_code/flica/data_std.mat"
#writeMat("./MATLAB_code/flica/data_std.mat",
#         mod1_std_cmplt = mod1_std_cmplt, mod2_std_cmplt = mod2_std_cmplt # complete cases from standardized missing data
#         )
# Run LICA on completers
#re_completer = flica_completer(ncomp = 5, flica_niter = 1500)